+++ front-end/src/App.css
@@ -16,7 +16,7 @@
 .App-header {
   /* From Example App */
   border: 1px solid black;
-  max-width: 500px;
+  max-width: 30rem;
   margin: 20px auto;
 
   background-color: #ffffff;
