+++ front-end/src/MyFridge.css
@@ -1,7 +1,9 @@
+/* List View Table Stuff */
 .header {
   text-align: left;
   padding-left: 40px;
   font-size: 18px;
+  font-weight: 400;
 }
 
 table {
@@ -33,50 +35,33 @@ td:nth-child(odd):hover {
   background-color: #ddd;
   /* color: white; */
 }
-/* table {
-  padding-top: 0px;
-  border-collapse: separate !important;
-  padding-left: 50px;
-  padding-right: 50px;
-} */
 
+/* Item Stuff */
 .title {
-  margin-right: 1rem;
-}
-
-.daysChip {
-  margin-right: 1rem;
+  margin-right: 0.6rem;
 }
 
 .chip {
-  margin-right: 1rem;
+  margin-right: 0.7rem;
   font-size: 1rem;
   font-weight: bold;
   display: inline;
-  border: 1px solid rgb(129, 205, 150);
   border-radius: 10px;
   padding-left: 0.3rem;
   padding-right: 0.5rem;
-  color: rgb(129, 205, 150);
 }
-.chip2 {
-  font-weight: 400;
+.chipFilled {
+  font-weight: 500;
   background-color: rgb(129, 205, 150);
   color: white;
 }
 
-/* p {
-  text-align: left;
-  padding-left: 50px;
-  font-size: 20px;
-} */
-
 .dot {
   height: 0.8rem;
-  width: 0.8rem;
-  border-radius: 50%;
+  width: 0.75rem;
+  border-radius: 100%;
   display: inline-block;
-  margin-right: 1rem;
+  margin-right: 0.8rem;
 }
 
 .dotGreen,
@@ -95,4 +80,19 @@ td:nth-child(odd):hover {
 .chipRed {
   background-color: rgb(205, 129, 129);
   border: 1px solid rgb(205, 129, 129);
+}
+
+.chipLots {
+  border: 1px solid rgb(129, 205, 150);
+  color: rgb(129, 205, 150);
+}
+
+.chipSome {
+  border: 1px solid rgb(200, 205, 129);
+  color: rgb(200, 205, 129);
+}
+
+.chipFew {
+  border: 1px solid rgb(205, 129, 129);
+  color: rgb(205, 129, 129);
 }
